pciaccess: make linux rom reading fallback optional
On kernels which have ROM files, the boot VGA rom file
will contain the 0xc0000 file anyways. If another card
is missing a rom file there is no need for this.

This is fallback is disabled by default, on the premise
that distro shipping this new a pciaccess will be running
on kernels with working rom support which is been upstream
for a long time