Fix long standing MTRR bug.
Add an include that makes the file common_vgaarb.c see the same
struct pci_system as the file linux_sysfs.c.

Without this fix, on a MTRR system, the vgaarb_fd member
would overwrite the mtrr_fd member and cause the MTRR_IOC_ADD_ENTRY call
to be issued to the incorrect device causing the infamous
"Inappropriate ioctl for device (25)" error.

This error would cause MTRR setup to fail on all systems relying on it
and is severe enough to warrant a new release.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Signed-off-by: Dave Airlie <airlied@redhat.com>