config: use LDADD to obtain appropriate dependencies
In generated scanpci/Makefile, this line is added:
scanpci_DEPENDENCIES = $(top_builddir)/src/libpciaccess.la

If the content of the library changes, the tool will rebuild.
When scanpci was in the same directory as the library, the .la file was used.
This will also ensure the "installed" pciaccess library is not used.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>